backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
celery[redis]>=5.3.4
redis>=4.5.0
pydantic>=1.10.13,<2.0
python-dotenv>=1.0.0
SQLAlchemy[asyncio]>=2.0.23
asyncpg>=0.28.0
aiosqlite>=0.19.0
alembic>=1.12.0
httpx>=0.25.0
pillow>=10.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
deep-translator>=1.11.4
